numpy.float is deprecated. Replace with python's float.

+++ bliss/common/axis.py
@@ -2195,14 +2195,14 @@ class Axis(Scannable):
 
     @lazy_init
     def set_event_positions(self, positions):
-        dial_positions = self.user2dial(numpy.array(positions, dtype=numpy.float))
+        dial_positions = self.user2dial(numpy.array(positions, dtype=float))
         step_positions = dial_positions * self.steps_per_unit
         return self.__controller.set_event_positions(self, step_positions)
 
     @lazy_init
     def get_event_positions(self):
         step_positions = numpy.array(
-            self.__controller.get_event_positions(self), dtype=numpy.float
+            self.__controller.get_event_positions(self), dtype=float
         )
         dial_positions = self.dial2user(step_positions)
         return dial_positions / self.steps_per_unit
